Changed timeout in msec from double to long Added TTimerActionAwaiter Fixed /// glitches

+++ BLTools Dot Net Std 2.0/Actions and functions/Awaiters and monitors/AConditionAwaiter.cs
@@ -7,9 +7,10 @@ namespace BLTools
 {
     public abstract class AConditionAwaiter : IConditionAwaiter
     {
-        protected double _TimeoutInMsec;
+        protected long _TimeoutInMsec;
 
         protected int _DurationToAwait;
         protected int _AwaitedDuration;
+
     }
 }
